+++ lib/bottom_navy_bar.dart
@@ -3,148 +3,166 @@ library bottom_navy_bar;
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 
-class BottomNavyBar extends StatefulWidget {
-  int selectedIndex;
+class BottomNavyBar extends StatelessWidget {
+  final int selectedIndex;
   final double iconSize;
   final Color backgroundColor;
   final bool showElevation;
+  final Duration animationDuration;
   final List<BottomNavyBarItem> items;
   final ValueChanged<int> onItemSelected;
-  final bool limitedNumberOfItems;
-  final int selectedWidth;
-  final int unselectedWidth;
+  final MainAxisAlignment mainAxisAlignment;
+  final double itemCornerRadius;
+  final double containerHeight;
+  final Curve curve;
 
-  BottomNavyBar(
-      {Key key,
-      this.selectedIndex = 0,
-      this.showElevation = true,
-      this.iconSize = 24,
-      this.backgroundColor,
-      this.limitedNumberOfItems = true,
-      this.selectedWidth = 130,
-      this.unselectedWidth = 50,
-      @required this.items,
-      @required this.onItemSelected}) {
+  BottomNavyBar({
+    Key key,
+    this.selectedIndex = 0,
+    this.showElevation = true,
+    this.iconSize = 24,
+    this.backgroundColor,
+    this.itemCornerRadius = 50,
+    this.containerHeight = 56,
+    this.animationDuration = const Duration(milliseconds: 270),
+    this.mainAxisAlignment = MainAxisAlignment.spaceBetween,
+    @required this.items,
+    @required this.onItemSelected,
+    this.curve = Curves.linear,
+  }) {
     assert(items != null);
-    assert(items.length >= 2 && (items.length <= 5 || !limitedNumberOfItems));
+    assert(items.length >= 2 && items.length <= 5);
     assert(onItemSelected != null);
-  }
-
-  @override
-  _BottomNavyBarState createState() {
-    return _BottomNavyBarState(
-        items: items,
-        backgroundColor: backgroundColor,
-        iconSize: iconSize,
-        onItemSelected: onItemSelected,
-        limitedNumberOfItems: limitedNumberOfItems,
-        selectedWidth: selectedWidth,
-        unselectedWidth: unselectedWidth);
-  }
-}
-
-class _BottomNavyBarState extends State<BottomNavyBar> {
-  final double iconSize;
-  Color backgroundColor;
-  List<BottomNavyBarItem> items;
-
-  ValueChanged<int> onItemSelected;
-  final bool limitedNumberOfItems;
-  final int selectedWidth;
-  final int unselectedWidth;
-
-  @override
-  void initState() {
-    super.initState();
-  }
-
-  _BottomNavyBarState(
-      {@required this.items,
-      this.backgroundColor,
-      this.iconSize,
-      @required this.onItemSelected,
-      this.limitedNumberOfItems,
-      this.selectedWidth,
-      this.unselectedWidth});
-
-  Widget _buildItem(BottomNavyBarItem item, bool isSelected) {
-    return AnimatedContainer(
-      width: isSelected
-          ? this.selectedWidth.toDouble()
-          : this.unselectedWidth.toDouble(),
-      height: double.maxFinite,
-      duration: Duration(milliseconds: 270),
-      padding: EdgeInsets.only(left: 12),
-      decoration: BoxDecoration(
-        color: isSelected ? item.activeColor.withOpacity(0.2) : backgroundColor,
-        borderRadius: BorderRadius.all(Radius.circular(50)),
-      ),
-      child: ListView(
-        shrinkWrap: true,
-        physics: NeverScrollableScrollPhysics(),
-        scrollDirection: Axis.horizontal,
-        children: <Widget>[
-          Row(
-            mainAxisAlignment: MainAxisAlignment.start,
-            crossAxisAlignment: CrossAxisAlignment.center,
-            children: <Widget>[
-              Padding(
-                padding: const EdgeInsets.only(right: 8),
-                child: IconTheme(
-                  data: IconThemeData(
-                      size: iconSize,
-                      color: isSelected
-                          ? item.activeColor.withOpacity(1)
-                          : item.inactiveColor == null
-                              ? item.activeColor
-                              : item.inactiveColor),
-                  child: item.icon,
-                ),
-              ),
-              isSelected
-                  ? DefaultTextStyle.merge(
-                      style: TextStyle(
-                          color: item.activeColor, fontWeight: FontWeight.bold),
-                      child: item.title,
-                    )
-                  : SizedBox.shrink()
-            ],
-          )
-        ],
-      ),
-    );
+    assert(curve != null);
   }
 
   @override
   Widget build(BuildContext context) {
-    backgroundColor = (backgroundColor == null)
+    final bgColor = (backgroundColor == null)
         ? Theme.of(context).bottomAppBarColor
         : backgroundColor;
 
     return Container(
-      decoration: BoxDecoration(color: backgroundColor, boxShadow: [
-        if (widget.showElevation)
-          BoxShadow(color: Colors.black12, blurRadius: 2)
-      ]),
+      decoration: BoxDecoration(
+        color: bgColor,
+        boxShadow: [
+          if (showElevation)
+            const BoxShadow(
+              color: Colors.black12,
+              blurRadius: 2,
+            ),
+        ],
+      ),
       child: SafeArea(
         child: Container(
           width: double.infinity,
-          height: 56,
-          padding: EdgeInsets.only(left: 8, right: 8, top: 6, bottom: 6),
+          height: containerHeight,
+          padding: const EdgeInsets.symmetric(vertical: 6, horizontal: 8),
           child: Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
+            mainAxisAlignment: mainAxisAlignment,
             children: items.map((item) {
               var index = items.indexOf(item);
               return GestureDetector(
-                onTap: () {
-                  onItemSelected(index);
-                  setState(() {
-                    widget.selectedIndex = index;
-                  });
-                },
-                child: _buildItem(item, widget.selectedIndex == index),
+                onTap: () => onItemSelected(index),
+                child: _ItemWidget(
+                  item: item,
+                  iconSize: iconSize,
+                  isSelected: index == selectedIndex,
+                  backgroundColor: bgColor,
+                  itemCornerRadius: itemCornerRadius,
+                  animationDuration: animationDuration,
+                  curve: curve,
+                ),
               );
             }).toList(),
+          ),
+        ),
+      ),
+    );
+  }
+}
+
+class _ItemWidget extends StatelessWidget {
+  final double iconSize;
+  final bool isSelected;
+  final BottomNavyBarItem item;
+  final Color backgroundColor;
+  final double itemCornerRadius;
+  final Duration animationDuration;
+  final Curve curve;
+
+  const _ItemWidget({
+    Key key,
+    @required this.item,
+    @required this.isSelected,
+    @required this.backgroundColor,
+    @required this.animationDuration,
+    @required this.itemCornerRadius,
+    @required this.iconSize,
+    this.curve = Curves.linear,
+  })  : assert(isSelected != null),
+        assert(item != null),
+        assert(backgroundColor != null),
+        assert(animationDuration != null),
+        assert(itemCornerRadius != null),
+        assert(iconSize != null),
+        assert(curve != null),
+        super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Semantics(
+      container: true,
+      selected: isSelected,
+      child: AnimatedContainer(
+        width: isSelected ? 130 : 50,
+        height: double.maxFinite,
+        duration: animationDuration,
+        curve: curve,
+        decoration: BoxDecoration(
+          color:
+              isSelected ? item.activeColor.withOpacity(0.2) : backgroundColor,
+          borderRadius: BorderRadius.circular(itemCornerRadius),
+        ),
+        child: SingleChildScrollView(
+          scrollDirection: Axis.horizontal,
+          physics: NeverScrollableScrollPhysics(),
+          child: Container(
+            width: isSelected ? 130 : 50,
+            padding: EdgeInsets.symmetric(horizontal: 8),
+            child: Row(
+              mainAxisSize: MainAxisSize.max,
+              mainAxisAlignment: MainAxisAlignment.start,
+              crossAxisAlignment: CrossAxisAlignment.center,
+              children: <Widget>[
+                IconTheme(
+                  data: IconThemeData(
+                    size: iconSize,
+                    color: isSelected
+                        ? item.activeColor.withOpacity(1)
+                        : item.inactiveColor == null
+                            ? item.activeColor
+                            : item.inactiveColor,
+                  ),
+                  child: item.icon,
+                ),
+                if (isSelected)
+                  Expanded(
+                    child: Container(
+                      padding: EdgeInsets.symmetric(horizontal: 4),
+                      child: DefaultTextStyle.merge(
+                        style: TextStyle(
+                          color: item.activeColor,
+                          fontWeight: FontWeight.bold,
+                        ),
+                        maxLines: 1,
+                        textAlign: item.textAlign,
+                        child: item.title,
+                      ),
+                    ),
+                  ),
+              ],
+            ),
           ),
         ),
       ),
@@ -157,12 +175,15 @@ class BottomNavyBarItem {
   final Text title;
   final Color activeColor;
   final Color inactiveColor;
+  final TextAlign textAlign;
 
-  BottomNavyBarItem(
-      {@required this.icon,
-      @required this.title,
-      this.activeColor = Colors.blue,
-      this.inactiveColor}) {
+  BottomNavyBarItem({
+    @required this.icon,
+    @required this.title,
+    this.activeColor = Colors.blue,
+    this.textAlign,
+    this.inactiveColor,
+  }) {
     assert(icon != null);
     assert(title != null);
   }