+++ CHANGELOG.md
@@ -1,25 +1,55 @@
-## [5.1.0] - Added assertion to prevent more than five items
+## 5.4.0
 
-## [5.0.0] - Fixed: Bar is too low at iPhone X(s/R)
+* Added `containerHeight` param to change the Navigation Bar height
 
-## [4.2.0] - Fixed the width of the Bottom navigation bar
+## 5.3.2
 
-## [4.1.0] - Added support to hide/show appBar's elevation
+* Added `curve` param to customize the item change's animation
+* Added `textAlign` property in `BottomNavyBarItem` to change the alignment of the item title
+ 
+## 5.2.2
+ * Added [mainAxisAlignment] property to align the items in the navigation bar
 
-## [4.0.0] - Fixed: Can't set default active item before Tap
+## 5.2.1 
+* Added [itemCornerRadius] property to change the active item's corner radius
 
-## [3.0.0] - animation duration set to 270 milliseconds && code improved.
+## 5.2.0 
+* Added optional parameter animationDuration to BottomNavyBar
+* Fixed issue, that selectedIndex has to be final in StatefulWidget
+* BottomNavyBar changed to StatelessWidget
 
-## [2.2.0] - icon now have default inactive color as the active color (bug fixed).
+## 5.1.0 
+* Added assertion to prevent more than five items
 
-## [2.1.0] - icon now have default inactive color as the active color.
+## 5.0.0 
+* Fixed: Bar is too low at iPhone X(s/R)
 
-## [2.0.0] - Full update & more improved.
+## 4.2.0 
+* Fixed the width of the Bottom navigation bar
 
-## [1.1.1] - First stable version.
+## 4.1.0 
+* Added support to hide/show appBar's elevation
 
-## [0.1.1] - Few updates.
+## 4.0.0 
+* Fixed: Can't set default active item before Tap
 
-## [0.1.0] - Initial beta release.
+## 3.0.0 
+* Animation duration set to 270 milliseconds && code improved.
 
-* Beta release.
+## 2.2.0 
+* Icon now have default inactive color as the active color (bug fixed).
+
+## 2.1.0 
+* Icon now have default inactive color as the active color.
+
+## 2.0.0 
+* Full update & more improved.
+
+## 1.1.1 
+* First stable version.
+
+## 0.1.1 
+* Few updates.
+
+## 0.1.0 
+* Initial beta release.